more

+++ inst/examples/randombits.c
@@ -3,7 +3,7 @@ SEXP C_randombits(SEXP n)
     if (!Rf_isInteger(n) || XLENGTH(n) != 1)
         Rf_error("`n` should be a single integer");
 
-    size_t _n = INTEGER(n)[0];
+    int _n = INTEGER(n)[0];
     if (_n == NA_INTEGER || _n < 1)
         Rf_error("incorrect `n`");
 
@@ -11,8 +11,8 @@ SEXP C_randombits(SEXP n)
     int* yp = INTEGER(y);
 
     GetRNGstate();
-    for (size_t i=0; i<_n; ++i)
-        yp[i] = (int)(unif_rand()>0.5);
+    for (int i=0; i<_n; ++i)
+        yp[i] = (int)(unif_rand()>0.5);  // not the best way to sample bits
     PutRNGstate();
 
     UNPROTECT(1);
